Remove unnecessary type casts of printf arguments to unsigned long (code refactoring)
* alloc.c [!NO_DEBUGGING] (GC_print_heap_sects): Remove unnecessary
cast to unsigned long in GC_*_printf call and adjust printf format
specifier to match the argument type.
* checksums.c [CHECKSUMS] (GC_checksum): Likewise.
* mark.c [PARALLEL_MARK] (GC_mark_local): Likewise.
* tests/test.c (typed_test): Likewise.
* checksums.c [CHECKSUMS] (GC_checksum): Change type of GC_n_clean,
GC_n_dirty from int to unsigned long.

+++ alloc.c
@@ -1206,9 +1206,9 @@ GC_INNER void GC_add_to_heap(struct hblk *p, size_t bytes)
       for (h = (struct hblk *)start; (word)h < (word)(start + len); h++) {
         if (GC_is_black_listed(h, HBLKSIZE)) nbl++;
       }
-      GC_printf("Section %d from %p to %p %lu/%lu blacklisted\n",
+      GC_printf("Section %d from %p to %p %u/%lu blacklisted\n",
                 i, (void *)start, (void *)&start[len],
-                (unsigned long)nbl, (unsigned long)divHBLKSZ(len));
+                nbl, (unsigned long)divHBLKSZ(len));
     }
   }
 #endif
